+++ lib/openapi3_validator.rb
@@ -110,10 +110,21 @@ class Openapi3Validator
   end
 
   def self.validate(req, res)
+    # prepare specs
     meth      = req.request_method.downcase
     path_spec = Openapi3Validator.spec.paths.match(req.path) || raise(Errors::PathNotFound, "Can't find path spec for #{meth} #{req.path}")
     meth_spec = path_spec.public_send(meth) || raise(Errors::MethodNotFound, "Can't find method spec for #{meth} #{req.path}")
     resp_spec = (meth_spec.responses.find { |k, _| k == res.status.to_s } || meth_spec.responses.find { |k, _| k == 'default' })&.last || raise(Errors::StatusNotFound, "Can't find matching status in spec: #{meth} #{req.path} -> #{res.status}")
+    req_spec = meth_spec.request_body
+
+    # validate request
+    if req_spec
+      require 'pp'
+      pp req
+      exit
+    end
+
+    # content empty?
     if resp_spec.content.to_a.empty?
       if res.body.size.zero?
         return
@@ -121,6 +132,8 @@ class Openapi3Validator
         raise(Errors::ExpectedNoContent, "#{meth} #{req.path} -> #{res.status}\nGot body: #{res.body.inspect}")
       end
     end
+
+    # find schema
     type = (res.headers['Content-Type'] || res.headers['content-type'])&.split(';')&.first
     if !type.nil? && resp_spec.content[type].nil?
       raise(Errors::UnexpectedContentType, "#{meth} #{req.path} -> #{res.status} unexpected content type #{type}")
@@ -128,6 +141,8 @@ class Openapi3Validator
     content = resp_spec.content[type]
     schema = content&.schema&.to_h
     return unless schema
+
+    # validate response
     begin
       JSON::Validator.validate!(schema, res.body)
     rescue JSON::Schema::ValidationError => e