+++ spec/spec_helper.rb
@@ -16,6 +16,7 @@
 
 require 'rack/test'
 require 'openapi3_validator/rspec/matchers'
+require_relative './support/test_app.rb'
 
 Openapi3Validator.config.spec = File.expand_path('../support/openapi.yml', __FILE__)