+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * Initial import
 * Fix problem with default status
 * Raise `SchemaValidationFailed` exception whether body is broken
+* Add request body validation